requirements.txt
openpyxl==3.1.5
python-dotenv==1.0.1
pyyaml==6.0.2
Pillow==11.2.1
